Implementing overwrite toggle as an attribute

+++ providers/record.rb
@@ -18,6 +18,10 @@ action :create do
     @ttl ||= new_resource.ttl
   end
 
+  def overwrite
+    @overwrite ||= new_resource.overwrite
+  end
+
   def zone
     @zone ||= Fog::DNS.new({ :provider => "aws",
                              :aws_access_key_id => new_resource.aws_access_key_id,
@@ -44,8 +48,13 @@ action :create do
     create
     Chef::Log.info "Record created: #{name}"
   elsif value != record.value.first
-    record.destroy
-    create
+    if overwrite
+      record.destroy
+      create
     Chef::Log.info "Record modified: #{name}"
+    else
+      Chef::Log.info "Record should have been modified but overwrite is disabled."
+    end
   end
+
 end
